Add files via upload
fix compiler warnings
clarify dotted rhythms in LAZY mode
remove debug code

+++ Composer.h
@@ -1,5 +1,5 @@
 //
-// TextMIDITools Version 1.0.83
+// TextMIDITools Version 1.0.84
 //
 // textmidicgm 1.0
 // Copyright © 2024 Thomas E. Janzen
@@ -43,7 +43,7 @@ namespace textmidi
             };
 
             Composer(bool gnuplot, bool answer, arrangements::PermutationEnum track_scramble,
-                TicksDuration track_scramble_period, int max_events_per_track = 100000)
+                TicksDuration track_scramble_period, size_t max_events_per_track = 100000)
               : gnuplot_(gnuplot),
                 answer_(answer),
                 track_scramble_(track_scramble, track_scramble_period),
@@ -72,7 +72,7 @@ namespace textmidi
             bool gnuplot_;
             bool answer_;
             TrackScramble track_scramble_;
-            int max_events_per_track_;
+            size_t max_events_per_track_;
             std::random_device random_dev_;
             std::mt19937 generator_;
         };
